Fix #12 - a test problem, not codebase problem
test_000 modified the Node class to test a case that
doesn't appear in the object model as such. For some reason,
even though Node class was imported explicitly for other
tests, when tests were run via tox, the monkey patch in
test_000 seemed to persist, leading to the error reported
in #12.

Refactored test_000 to create a separate test class FakeNode
to test the edge case. This seemed to clear up the test run issue
described in #12.